+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ master (unreleased)
 ===================
 
 - Trim whitespaces in the generated ``formidable.js`` file. This is more than just cosmetics, it prevents to have a polluted history on this file.
+- Hotfix: Extract conditions and filter them using the fields that exist in the form.
 
 Release 1.4.0 (2018-02-21)
 ==========================